trigger child window postMessage(session state) in more places where mass UI is launched
